Adds support for running `EVAL` with different scripting engines
