+++ src/eval.h
@@ -1,0 +1,8 @@
+#ifndef _EVAL_H_
+#define _EVAL_H_
+
+void evalInit(void);
+
+void *evalActiveDefragScript(void *ptr);
+
+#endif /* _EVAL_H_ */
