+++ src/lua/engine_lua.h
@@ -1,0 +1,9 @@
+#ifndef _ENGINE_LUA_
+#define _ENGINE_LUA_
+
+#include "../scripting_engine.h"
+#include <lua.h>
+
+int luaEngineInitEngine(void);
+
+#endif /* _ENGINE_LUA_ */
