+++ src/server.c
@@ -44,6 +44,9 @@
 #include "sds.h"
 #include "module.h"
 #include "scripting_engine.h"
+#include "lua/engine_lua.h"
+#include "lua/debug_lua.h"
+#include "eval.h"
 
 #include <time.h>
 #include <signal.h>
@@ -1366,6 +1369,13 @@ void checkChildrenDone(void) {
     }
 }
 
+static int sumEngineUsedMemory(scriptingEngine *engine, void *context) {
+    size_t *total_memory = (size_t *)context;
+    engineMemoryInfo mem_info = engineCallGetMemoryInfo(engine, VMSE_ALL);
+    *total_memory += mem_info.used_memory;
+    return 1;
+}
+
 /* Called from serverCron and cronUpdateMemoryStats to update cached memory metrics. */
 void cronUpdateMemoryStats(void) {
     /* Record the max memory used since the server was started. */
@@ -1391,8 +1401,9 @@ void cronUpdateMemoryStats(void) {
             /* LUA memory isn't part of zmalloc_used, but it is part of the process RSS,
              * so we must deduct it in order to be able to calculate correct
              * "allocator fragmentation" ratio */
-            size_t lua_memory = evalMemory();
-            server.cron_malloc_stats.allocator_resident = server.cron_malloc_stats.process_rss - lua_memory;
+            size_t engines_memory = 0;
+            engineManagerForEachEngine(sumEngineUsedMemory, &engines_memory);
+            server.cron_malloc_stats.allocator_resident = server.cron_malloc_stats.process_rss - engines_memory;
         }
         if (!server.cron_malloc_stats.allocator_active)
             server.cron_malloc_stats.allocator_active = server.cron_malloc_stats.allocator_resident;
@@ -2900,12 +2911,25 @@ void initServer(void) {
         serverPanic("Scripting engine manager initialization failed, check the server logs.");
     }
 
+    // Since we initialized the scripting engine manager, we need to ensure
+    // that commands with `CMD_NOSCRIPT` flag are not allowed to run in scripts.
+    //
+    server.script_disable_deny_script = 0;
+
     /* Initialize the LUA scripting engine. */
-    scriptingInit(1);
+    if (luaEngineInitEngine() != C_OK) {
+        serverPanic("Lua engine initialization failed, check the server logs.");
+        exit(1);
+    }
+
     /* Initialize the functions engine based off of LUA initialization. */
     if (functionsInit() == C_ERR) {
         serverPanic("Functions initialization failed, check the server logs.");
     }
+
+    /* Initialize the EVAL scripting component. */
+    evalInit();
+
     slowlogInit();
     latencyMonitorInit();
     initSharedQueryBuf();
