+++ tests/unit/functions.tcl
@@ -323,7 +323,7 @@ start_server {tags {"scripting repl external:skip"}} {
 
         test {FUNCTION - creation is replicated to replica} {
             r function load [get_no_writes_function_code LUA test test {return 'hello'}]
-            wait_for_condition 150 100 {    
+            wait_for_condition 150 100 {
                 [r -1 function list] eq {{library_name test engine LUA functions {{name test description {} flags no-writes}}}}
             } else {
                 fail "Failed waiting for function to replicate to replica"
@@ -1062,7 +1062,7 @@ start_server {tags {"scripting"}} {
 
     test {FUNCTION - deny oom} {
         r FUNCTION load replace {#!lua name=test
-            server.register_function('f1', function() return redis.call('set', 'x', '1') end) 
+            server.register_function('f1', function() return redis.call('set', 'x', '1') end)
         }
 
         r config set maxmemory 1
@@ -1093,7 +1093,7 @@ start_server {tags {"scripting"}} {
             server.register_function{function_name='f3', callback=function() return redis.call('get', 'x') end, flags={'allow-stale', 'no-writes'}}
             server.register_function{function_name='f4', callback=function() return redis.call('info', 'server') end, flags={'allow-stale', 'no-writes'}}
         }
-        
+
         r config set replica-serve-stale-data no
         r replicaof 127.0.0.1 1
 
@@ -1166,7 +1166,7 @@ start_server {tags {"scripting"}} {
             server.register_function('f3', function() return 1 end)
         }} e
         assert_match "*Library 'test1' already exists*" $e
-        
+
 
         r function stats
     } {running_script {} engines {LUA {libraries_count 1 functions_count 2}}}
@@ -1214,7 +1214,7 @@ start_server {tags {"scripting"}} {
         r FUNCTION FLUSH
 
         r FUNCTION load {#!lua name=test1
-            server.register_function('f1', function() 
+            server.register_function('f1', function()
                 mt = getmetatable(_G)
                 original_globals = mt.__index
                 original_globals['redis'] = function() return 1 end
